requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
procgen>=0.10.7
numpy>=1.24.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0

